finalized project

+++ su2nn_e3nn_core/nn/_fc.py
@@ -39,12 +39,12 @@ class _Layer(torch.nn.Module):
         # - PROFILER - with torch.autograd.profiler.record_function(self._profiling_str):
         if self.act is not None:
             w = self.weight / (self.h_in * self.var_in) ** 0.5
-            x = x @ w
+            x = x @ w.to(dtype = x.dtype)
             x = self.act(x)
             x = x * self.var_out**0.5
         else:
             w = self.weight / (self.h_in * self.var_in / self.var_out) ** 0.5
-            x = x @ w
+            x = x @ w.to(dtype = x.dtype)
         return x
